# v1/wwwroot/openapi.yaml
openapi: 3.0.3
info:
  title: ASP.NET Core | Entity Framework Core
  description: |-
    Данное Rest-API выполняет роль backend части для [NetheriteTrade](https://t.me/netheritetrade_bot/netheriteapp) и может быть использовано для
    1) Добавление задач
    2) Регистрации пользователей
    3) Майнинга виртуальной валюты
  version: 1.0.11
servers:
  - url: https://5.35.81.57
  - url: http://localhost:5068
...
paths:
    /tasks:
      post:
        tags:
          - Tasks
        summary: Добавить новую Task
        description: Добавление новой задачи Task
        operationId: CreateTask
        parameters:
          - name: title
            in: query
            description: Название задачи
            required: true
            schema:
              type: string
              example: "Подписаться на канал"
          - name: description
            in: query
            description: Описание задачи
            required: true
            schema:
              type: string
              example: "Перейдите по ссылке, оформите подписку и нажмите 'Готово'"
          - name: price
            in: query
            description: Награда за выполнение задачи
            required: true
            schema:
              type: string
              example: "100"
        responses:
          '200':
            description: Задача успешно добавлена
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    status:
                      type: integer
                      example: 200
                    callback:
                      type: object
                      properties:
                        id:
                          type: string
                          example: "f397aa41-cfa3-481a-b5c8-dec8de1f3ba6"
          '400':
            description: Отсутствуют необходимые параметры
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    status:
                      type: integer
                      example: 400
                    callback:
                      type: string
                      example: "Missing required parameters: title, description"
          '409':
            description: Price не является числом
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    status:
                      type: integer
                      example: 409
                    callback:
                      type: string
                      example: "Price must be a positive number"
                      
      get:
        tags:
          - Tasks
        summary: Список Tasks
        description: Получить полный список имеющихся Tasks задач
        operationId: GetTasks
        responses:
          '200':
            description: Коллекция задач успешно получена
            content:
              application/json:
                schema:
                  type: array
                  items:
                    type: object
                    properties:
                      id:
                        type: string
                        example: dbd5a353-583a-441f-8d26-0cd1d7924931
                      title:
                        type: string
                        example: Подписаться на канал
                      description:
                        type: string
                        example: Перейдите по ссылке, оформите подписку и нажмите 'Готово'
                      price:
                        type: string
                        example: 100
                      icon:
                        type: string
                        format: binary
                        example: iVBORw0KGgoAAAANSUhEUgAA...
...
    /mining/start:
      post:
        tags:
          - Mining
        summary: Запуск процесса майнинга
        description: Запускает процесс майнинга для пользователя с указанным UUID. Под капотом так же есть параметр earn, который в теории позволяет регулировать начисление монет за майнинг. Данный параметр используется в **POST /user/{telegram_id}/complete/{task_id}** На данный момент за 1 майнинг начисляется фиксированно 100 монет.
        operationId: StartMining
        parameters:
          - name: id
            in: query
            description: UUID пользователя
            required: true
            schema:
              type: string
              example: "f397aa41-cfa3-481a-b5c8-dec8de1f3ba6"
        responses:
          '200':
            description: Майнинг успешно запущен
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    status:
                      type: integer
                      example: 200
                    callback:
                      type: boolean
                      example: null
          '400':
            description: Недостает параметра id
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    status:
                      type: integer
                      example: 400
                    callback:
                      type: string
                      example: "Missing required parameter: id"
          '409':
            description: Юзер уже майнит. Нельзя начать майнить дважды.
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    status:
                      type: integer
                      example: 409
                    callback:
                      type: string
                      example: "Miner with this id already exists"
...